+++ NEWS.md
@@ -1,10 +1,12 @@
 # Version 2.1.0.9000
 * Removed class restrictions on objects passed to `regTable()`. `regTable()`
   now works with almost every kind of regression object.
+* Added `[.regTable` method for intelligent subsetting of regTable objects (#27).
 * Added global option `Bullock.print.regTable.dp` on package load so that 
   users can change default number of digits that are displayed when they
   `print` a regTable object.
 * Added `envir` argument to `lsos()` and `.ls.objects()` (#6).
+* Added str.regTable().
 * Fixed bug in latexTable footers when `SE_table` is `FALSE` (#24).
 * Improved vignette formatting (#19).  
 &nbsp;